test: Clean up and expand test coverage for MCP route extensions
- Removed `AbstractKtorExtensionsTest` and migrated its helpers to `TestHelpers.kt` for better reuse.
- Added new test cases for route and application-level MCP extensions, ensuring default and custom paths work as intended.
- Enabled parallel test execution via `junit-platform.properties`.
- Updated (reduced) detekt baselines to reflect changes in test structure and layout.
